requirements.txt
Django>=5.0
djangorestframework
django-filter
drf-spectacular
django-cors-headers
Pillow
python-dotenv
psycopg2-binary
whitenoise
django-jazzmin
